Updates header
